requirements.txt
flask==3.0.2
yt-dlp==2025.06.09
python-dotenv==1.0.1
gunicorn==21.2.0 